Extract hash calculation
