+++ components/telemetry-operator/internal/hash/hash_test.go
@@ -1,4 +1,4 @@
-package tracepipeline
+package hash
 
 import (
 	"testing"
@@ -41,18 +41,18 @@ var (
 )
 
 func TestEqualConfig(t *testing.T) {
-	hash1 := CreateConfigHash([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
-	hash2 := CreateConfigHash([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
+	hash1 := Calculate([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
+	hash2 := Calculate([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
 	require.Equal(t, hash1, hash2)
 }
 
 func TestUnequalConfig(t *testing.T) {
-	hash1 := CreateConfigHash([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
-	hash2 := CreateConfigHash([]corev1.ConfigMap{configMap2}, []corev1.Secret{secret2})
+	hash1 := Calculate([]corev1.ConfigMap{configMap1}, []corev1.Secret{secret1})
+	hash2 := Calculate([]corev1.ConfigMap{configMap2}, []corev1.Secret{secret2})
 	require.NotEqual(t, hash1, hash2)
 }
 
 func TestEmptyConfig(t *testing.T) {
-	hash := CreateConfigHash([]corev1.ConfigMap{emptyConfigMap}, []corev1.Secret{emptySecret})
+	hash := Calculate([]corev1.ConfigMap{emptyConfigMap}, []corev1.Secret{emptySecret})
 	require.NotEmpty(t, hash)
 }
